requirements.txt
fastapi
uvicorn
pydantic
pytest
python-dotenv
flake8
pre-commit
black